82. Clean Up API Urls and Views
