+++ static/js/main.943a2867.chunk.js
@@ -1,0 +1,2 @@
+(this["webpackJsonptweetme2-web"]=this["webpackJsonptweetme2-web"]||[]).push([[0],{14:function(e,t,n){},15:function(e,t,n){},17:function(e,t,n){"use strict";n.r(t);var c=n(1),a=n.n(c),s=n(6),r=n.n(s),o=(n(14),n.p+"static/media/logo.6ce24c58.svg");n(15);function i(e,t,n,c){var a;c&&(a=JSON.stringify(c));var s=new XMLHttpRequest,r="http://localhost:8000/api".concat(t);s.responseType="json";var o=function(e){var t=null;if(document.cookie&&""!==document.cookie)for(var n=document.cookie.split(";"),c=0;c<n.length;c++){var a=n[c].trim();if(a.substring(0,e.length+1)===e+"="){t=decodeURIComponent(a.substring(e.length+1));break}}return t}("csrftoken");s.open(e,r),s.setRequestHeader("Content-Type","application/json"),o&&(s.setRequestHeader("HTTP_X_REQUESTED_WITH","XMLHttpRequest"),s.setRequestHeader("X-Requested-With","XMLHttpRequest"),s.setRequestHeader("X-CSRFToken",o)),s.onload=function(){n(s.response,s.status)},s.onerror=function(e){console.log(e),n({message:"The request was an error"},400)},s.send(a)}var l=n(0);function u(e){var t=e.tweet,n=e.action,c=e.didPerformAction,a=t.likes?t.likes:0,s=e.className?e.className:"btn btn-primary btn-sm",r=n.display?n.display:"Action",o=function(e,t){console.log(e,t),200!==t&&201!==t||!c||c(e,t)},u="like"===n.type?"".concat(a," ").concat(r):r;return Object(l.jsx)("button",{className:s,onClick:function(e){e.preventDefault(),function(e,t,n){i("POST","/tweets/action/",n,{id:e,action:t})}(t.id,n.type,o)},children:u})}var d=n(9),j=n(3),b=n(2);function m(e){var t=a.a.createRef(),n=e.didTweet,c=function(e,t){201===t?n(e):(console.log(e),alert("An error occured please try again"))};return Object(l.jsx)("div",{className:e.className,children:Object(l.jsxs)("form",{onSubmit:function(e){e.preventDefault();var n=t.current.value;i("POST","/tweets/create/",c,{content:n}),t.current.value=""},children:[Object(l.jsx)("textarea",{ref:t,required:!0,className:"form-control",name:"tweet"}),Object(l.jsx)("button",{type:"submit",className:"btn btn-primary my-3",children:"Tweet"})]})})}var f=n(8);function w(e){var t=e.tweet;return t.parent?Object(l.jsx)("div",{className:"row",children:Object(l.jsxs)("div",{className:"col-11 mx-auto p-3 border rounded",children:[Object(l.jsx)("p",{className:"mb-0 text-muted small",children:"Retweet"}),Object(l.jsx)(p,{hideActions:!0,className:" ",tweet:t.parent})]})}):null}function p(e){var t=e.tweet,n=e.didRetweet,s=e.hideActions,r=Object(c.useState)(e.tweet?e.tweet:null),o=Object(b.a)(r,2),i=o[0],d=o[1],j=e.className?e.className:"col-10 mx-auto col-md-6",m=window.location.pathname.match(Object(f.a)(/([0-9]+)/,{tweetid:1})),p=m?m.groups.tweetid:-1,O="".concat(t.id)==="".concat(p),h=function(e,t){200===t?d(e):201===t&&n&&n(e)};return Object(l.jsxs)("div",{className:j,children:[Object(l.jsxs)("div",{children:[Object(l.jsxs)("p",{children:[t.id," - ",t.content]}),Object(l.jsx)(w,{tweet:t})]}),Object(l.jsxs)("div",{className:"btn btn-group",children:[i&&!0!==s&&Object(l.jsxs)(a.a.Fragment,{children:[Object(l.jsx)(u,{tweet:i,didPerformAction:h,action:{type:"like",display:"Likes"}}),Object(l.jsx)(u,{tweet:i,didPerformAction:h,action:{type:"unlike",display:"Unlike"}}),Object(l.jsx)(u,{tweet:i,didPerformAction:h,action:{type:"retweet",display:"Retweet"}})]}),!0===O?null:Object(l.jsx)("button",{className:"btn btn-outline-primary btn-sm",onClick:function(e){e.preventDefault(),window.location.href="/".concat(t.id)},children:"View"})]})]})}function O(e){var t=Object(c.useState)([]),n=Object(b.a)(t,2),a=n[0],s=n[1],r=Object(c.useState)([]),o=Object(b.a)(r,2),u=o[0],d=o[1],m=Object(c.useState)(!1),f=Object(b.a)(m,2),w=f[0],O=f[1];Object(c.useEffect)((function(){var t=Object(j.a)(e.newTweets).concat(a);t.length!==u.length&&d(t)}),[e.newTweets,u,a]),Object(c.useEffect)((function(){if(!1===w){!function(e,t){var n="/tweets/";e&&(n="/tweets/?username=".concat(e)),i("GET",n,t)}(e.username,(function(e,t){200===t?(s(e),O(!0)):alert("There was an error")}))}}),[a,w,O,e.username]);var h=function(e){var t=Object(j.a)(a);t.unshift(e),s(t);var n=Object(j.a)(u);n.unshift(u),d(n)};return u.map((function(e,t){return Object(l.jsx)(p,{tweet:e,didRetweet:h,className:"my-5 py-5 border bg-white text-dark"},"".concat(t,"-{item.id}"))}))}function h(e){var t=Object(c.useState)([]),n=Object(b.a)(t,2),a=n[0],s=n[1],r="false"!==e.canTweet;return Object(l.jsxs)("div",{className:e.className,children:[!0===r&&Object(l.jsx)(m,{didTweet:function(e){var t=Object(j.a)(a);t.unshift(e),s(t)},className:"col-12 mb-3"}),Object(l.jsx)(O,Object(d.a)({newTweets:a},e))]})}function v(e){var t=e.tweetId,n=Object(c.useState)(!1),a=Object(b.a)(n,2),s=a[0],r=a[1],o=Object(c.useState)(null),u=Object(b.a)(o,2),d=u[0],j=u[1],m=function(e,t){200===t?j(e):alert("There was an error finding your tweet.")};return Object(c.useEffect)((function(){!1===s&&(!function(e,t){i("GET","/tweets/".concat(e,"/"),t)}(t,m),r(!0))}),[t,s,r]),null===d?null:Object(l.jsx)(p,{tweet:d,className:e.className})}var x=function(){return Object(l.jsx)("div",{className:"App",children:Object(l.jsxs)("header",{className:"App-header",children:[Object(l.jsx)("img",{src:o,className:"App-logo",alt:"logo"}),Object(l.jsxs)("p",{children:["Edit ",Object(l.jsx)("code",{children:"src/App.js"})," and save to reload."]}),Object(l.jsx)("div",{children:Object(l.jsx)(h,{})}),Object(l.jsx)("a",{className:"App-link",href:"https://reactjs.org",target:"_blank",rel:"noopener noreferrer",children:"Learn React"})]})})};Boolean("localhost"===window.location.hostname||"[::1]"===window.location.hostname||window.location.hostname.match(/^127(?:\.(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)){3}$/));var g=document.getElementById("root");g&&r.a.render(Object(l.jsx)(x,{}),g);var N=a.a.createElement,y=document.getElementById("tweetme-2");y&&r.a.render(N(h,y.dataset),y),document.querySelectorAll(".tweetme-2-detail").forEach((function(e){r.a.render(N(v,e.dataset),e)})),"serviceWorker"in navigator&&navigator.serviceWorker.ready.then((function(e){e.unregister()}))}},[[17,1,2]]]);
+//# sourceMappingURL=main.943a2867.chunk.js.map
