+++ templates/react/js.html
@@ -1,2 +1,2 @@
-<script src="/static/js/2.671dfcc9.chunk.js"></script>
-<script src="/static/js/main.4dcdbe38.chunk.js"></script>
+<script src="/static/js/2.98a64f88.chunk.js"></script>
+<script src="/static/js/main.943a2867.chunk.js"></script>
