+++ templates/tweets/detail.html
@@ -1,0 +1,10 @@
+{% extends 'base.html' %}
+
+{% block content %}
+<div 
+    class='tweetme-2-detail' 
+    data-tweet-id="{{tweet_id}}" 
+    data-class-name='col-6 mx-auto'></div>
+
+
+{% endblock content %}
