+++ templates/tweets/list.html
@@ -1,0 +1,7 @@
+{% extends 'base.html' %}
+
+{% block content %}
+
+<div id='tweetme-2'></div>
+
+{% endblock content %}
