+++ templates/tweets/profile.html
@@ -1,0 +1,9 @@
+{% extends 'base.html' %}
+
+{% block content %}
+
+    <div id="tweetme-2" 
+    data-username="{{ profile_username }}" 
+    data-can-tweet="false"></div>
+
+{% endblock content %}
