+++ tweetme2-web/src/index.js
@@ -15,12 +15,11 @@ if (tweetsEl) {
   ReactDOM.render(e(TweetsComponent, tweetsEl.dataset), tweetsEl);
 }
 
-const tweetDetailElements = document.querySelectorAll(".tweetme-2-detail")
+const tweetDetailElements = document.querySelectorAll(".tweetme-2-detail");
 
-tweetDetailElements.forEach(container=>{
-  ReactDOM.render(e(TweetDetailComponent, container.dataset), container)
-})
-
+tweetDetailElements.forEach((container) => {
+  ReactDOM.render(e(TweetDetailComponent, container.dataset), container);
+});
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
 // Learn more about service workers: https://bit.ly/CRA-PWA
