+++ tweetme2-web/src/tweets/detail.js
@@ -21,7 +21,15 @@ export function Tweet(props) {
   const className = props.className
     ? props.className
     : "col-10 mx-auto col-md-6";
+  const path = window.location.pathname;
+  const match = path.match(/(?<tweetid>\d+)/);
+  const urlTweetId = match ? match.groups.tweetid : -1;
+  const isDetail = `${tweet.id}` === `${urlTweetId}`;
 
+  const handleLink = (event) => {
+    event.preventDefault();
+    window.location.href = `/${tweet.id}`;
+  };
   const handlePerformAction = (newActionTweet, status) => {
     if (status === 200) {
       setActionTweet(newActionTweet);
@@ -40,25 +48,35 @@ export function Tweet(props) {
         </p>
         <ParentTweet tweet={tweet} />
       </div>
-      {actionTweet && hideActions !== true && (
-        <div className="btn btn-group">
-          <ActionBtn
-            tweet={actionTweet}
-            didPerformAction={handlePerformAction}
-            action={{ type: "like", display: "Likes" }}
-          />
-          <ActionBtn
-            tweet={actionTweet}
-            didPerformAction={handlePerformAction}
-            action={{ type: "unlike", display: "Unlike" }}
-          />
-          <ActionBtn
-            tweet={actionTweet}
-            didPerformAction={handlePerformAction}
-            action={{ type: "retweet", display: "Retweet" }}
-          />
-        </div>
-      )}
+      <div className="btn btn-group">
+        {actionTweet && hideActions !== true && (
+          <React.Fragment>
+            <ActionBtn
+              tweet={actionTweet}
+              didPerformAction={handlePerformAction}
+              action={{ type: "like", display: "Likes" }}
+            />
+            <ActionBtn
+              tweet={actionTweet}
+              didPerformAction={handlePerformAction}
+              action={{ type: "unlike", display: "Unlike" }}
+            />
+            <ActionBtn
+              tweet={actionTweet}
+              didPerformAction={handlePerformAction}
+              action={{ type: "retweet", display: "Retweet" }}
+            />
+          </React.Fragment>
+        )}
+        {isDetail === true ? null : (
+          <button
+            className="btn btn-outline-primary btn-sm"
+            onClick={handleLink}
+          >
+            View
+          </button>
+        )}
+      </div>
     </div>
   );
 }
